fix: Add a custom hook for the next action state handling
